+++ components/form-builder/hooks/useProgress.tsx
@@ -4,14 +4,26 @@ import { Responses } from "@lib/types";
 interface ProgressType {
   userProgress?: Responses | undefined;
   importProgress: Dispatch<SetStateAction<Responses | undefined>>;
+  saveProgress: (values: Responses | undefined) => void;
 }
 
 export const ProgressContext = createContext<ProgressType | null>(null);
 
 export function ProgressProvider({ children }: { children: React.ReactNode }) {
   const [userProgress, importProgress] = useState<Responses | undefined>({});
+
+  const saveProgress = (values: Responses | undefined) => {
+    const blob = new Blob([JSON.stringify(values)], { type: "application/json" });
+    const url = URL.createObjectURL(blob);
+    const a = document.createElement("a");
+    a.href = url;
+    a.download = "progress.json";
+    a.click();
+    URL.revokeObjectURL(url);
+  };
+
   return (
-    <ProgressContext.Provider value={{ userProgress, importProgress }}>
+    <ProgressContext.Provider value={{ userProgress, importProgress, saveProgress }}>
       {children}
     </ProgressContext.Provider>
   );
@@ -21,9 +33,8 @@ export const useProgress = () => {
   const value = useContext(ProgressContext);
   if (value === null) {
     return {
-      importProgress: () => {
-        return;
-      },
+      importProgress: () => null,
+      saveProgress: () => null,
       userProgress: undefined,
     };
   }
